vec is faster than my custom array type

+++ src/lib.rs
@@ -1,8 +1,8 @@
 //! The variables and literals are implemented as plain integers
 #[allow(non_camel_case_types)]
-pub type uint     = u32;
+pub type uint     = usize;
 #[allow(non_camel_case_types)]
-pub type iint     = i32;
+pub type iint     = isize;
 
 pub type Variable = uint;
 pub type Literal  = iint;
